Add another Asp.Net Core Web API project (#12)
With Asp.Net Core Web API project, it is much easier to host outside of Azure function. Like we have a development AKS clsuter that can run the contianer image in it. It is also easy to do local debug.

# Local debug & verification
There is a `Inclusive.HeatSensor.Web.http` file that store the testing http requests.

![image](https://github.com/user-attachments/assets/2f294dae-a2fc-4adf-9bf6-7ac204ace26d)